SNES: Make S9xMixSamplesLowPass call S9xMixSamples
This is functionally identical, as the old code did the clipping before applying the filter anyway.

It takes 150us more to go through the buffer a second time, to filter it. Which seems like a reasonable compromise for so much removed duplicated code.

+++ retro-core/components/snes9x/src/soundux.h
@@ -123,7 +123,7 @@ void S9xSetEchoFeedback(int32_t echo_feedback);
 void S9xSetEchoEnable(uint8_t byte);
 void S9xSetEchoDelay(int32_t byte);
 void S9xSetEchoWriteEnable(uint8_t byte);
-void S9xSetFilterCoefficient(int32_t tap, int32_t value);
+void S9xSetFilterCoefficient(int32_t tap, int8_t value);
 void S9xSetFrequencyModulationEnable(uint8_t byte);
 void S9xSetEnvelopeRate(int32_t channel, uint32_t rate, int32_t direction, int32_t target, uint32_t  mode);
 bool S9xSetSoundMode(int32_t channel, int32_t mode);
